move getResolveAndSubscriptionMethods to a separate file
